# backend/temp_openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/users/":{"get":{"tags":["Users"],"summary":"List Users","operationId":"list_users_users__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/UserResponse"},"type":"array","title":"Response List Users Users  Get"}}}}}},"post":{"tags":["Users"],"summary":"Register User","operationId":"register_user_users__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{user_id}":{"get":{"tags":["Users"],"summary":"Fetch User","operationId":"fetch_user_users__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Users"],"summary":"Update User","operationId":"update_user_users__user_id__put","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Users"],"summary":"Delete User","operationId":"delete_user_users__user_id__delete","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/questionnaire/create":{"post":{"tags":["Questionnaire"],"summary":"Create Question","operationId":"create_question_questionnaire_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/questionnaire/":{"get":{"tags":["Questionnaire"],"summary":"List Questions","operationId":"list_questions_questionnaire__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/QuestionResponse"},"type":"array","title":"Response List Questions Questionnaire  Get"}}}}}}},"/questionnaire/{q_id}":{"get":{"tags":["Questionnaire"],"summary":"Get Question","operationId":"get_question_questionnaire__q_id__get","parameters":[{"name":"q_id","in":"path","required":true,"schema":{"type":"string","title":"Q Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Questionnaire"],"summary":"Update Question","operationId":"update_question_questionnaire__q_id__patch","parameters":[{"name":"q_id","in":"path","required":true,"schema":{"type":"string","title":"Q Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Questionnaire"],"summary":"Remove Question","operationId":"remove_question_questionnaire__q_id__delete","parameters":[{"name":"q_id","in":"path","required":true,"schema":{"type":"string","title":"Q Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/response/submit":{"post":{"tags":["Response"],"summary":"Submit User Responses","operationId":"submit_user_responses_response_submit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitResponse"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/response/{user_id}":{"get":{"tags":["Response"],"summary":"Get User Responses","operationId":"get_user_responses_response__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponsesOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/predict/{user_id}":{"post":{"tags":["Predictions"],"summary":"Get Prediction","operationId":"get_prediction_predict__user_id__post","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/predict/latest/{user_id}":{"get":{"tags":["Predictions"],"summary":"Get Latest Prediction","operationId":"get_latest_prediction_predict_latest__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/predict/validate/{user_id}":{"get":{"tags":["Predictions"],"summary":"Validate Feature Coverage","operationId":"validate_feature_coverage_predict_validate__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"FeatureMapping":{"properties":{"feature_name":{"type":"string","title":"Feature Name"},"feature_value":{"type":"number","title":"Feature Value","default":1.0}},"type":"object","required":["feature_name"],"title":"FeatureMapping"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OptionCreate":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"text":{"type":"string","title":"Text"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"mappings":{"items":{"$ref":"#/components/schemas/FeatureMapping"},"type":"array","title":"Mappings"}},"type":"object","required":["text","mappings"],"title":"OptionCreate"},"OptionResponse":{"properties":{"id":{"type":"string","title":"Id"},"text":{"type":"string","title":"Text"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"mappings":{"items":{"$ref":"#/components/schemas/FeatureMapping"},"type":"array","title":"Mappings"}},"type":"object","required":["id","text","mappings"],"title":"OptionResponse"},"QuestionCreate":{"properties":{"id":{"type":"string","title":"Id"},"text":{"type":"string","title":"Text"},"category":{"type":"string","title":"Category"},"q_type":{"type":"string","enum":["yes_no","single_select","multi_select","input"],"title":"Q Type"},"is_initial":{"type":"boolean","title":"Is Initial","default":true},"priority":{"type":"integer","title":"Priority","default":0},"options":{"anyOf":[{"items":{"$ref":"#/components/schemas/OptionCreate"},"type":"array"},{"type":"null"}],"title":"Options"},"direct_mappings":{"anyOf":[{"items":{"$ref":"#/components/schemas/FeatureMapping"},"type":"array"},{"type":"null"}],"title":"Direct Mappings"}},"type":"object","required":["id","text","category","q_type"],"title":"QuestionCreate"},"QuestionResponse":{"properties":{"id":{"type":"string","title":"Id"},"text":{"type":"string","title":"Text"},"category":{"type":"string","title":"Category"},"q_type":{"type":"string","title":"Q Type"},"is_initial":{"type":"boolean","title":"Is Initial"},"priority":{"type":"integer","title":"Priority"},"options":{"anyOf":[{"items":{"$ref":"#/components/schemas/OptionResponse"},"type":"array"},{"type":"null"}],"title":"Options","default":[]}},"type":"object","required":["id","text","category","q_type","is_initial","priority"],"title":"QuestionResponse"},"QuestionUpdate":{"properties":{"text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Text"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"priority":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Priority"},"is_initial":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Initial"}},"type":"object","title":"QuestionUpdate"},"SingleResponse":{"properties":{"question_id":{"type":"string","title":"Question Id"},"selected_option_ids":{"items":{"type":"string"},"type":"array","title":"Selected Option Ids"}},"type":"object","required":["question_id","selected_option_ids"],"title":"SingleResponse"},"SingleResponseOut":{"properties":{"question_id":{"type":"string","title":"Question Id"},"selected_option_ids":{"items":{"type":"string"},"type":"array","title":"Selected Option Ids"}},"type":"object","required":["question_id","selected_option_ids"],"title":"SingleResponseOut"},"SubmitResponse":{"properties":{"user_id":{"type":"string","title":"User Id"},"responses":{"items":{"$ref":"#/components/schemas/SingleResponse"},"type":"array","title":"Responses"}},"type":"object","required":["user_id","responses"],"title":"SubmitResponse"},"UserCreate":{"properties":{"user_id":{"type":"string","title":"User Id"},"age":{"type":"integer","exclusiveMinimum":0.0,"title":"Age","description":"Age must be a positive integer"},"height":{"type":"number","exclusiveMinimum":0.0,"title":"Height","description":"Height must be a positive number in cm"},"weight":{"type":"number","exclusiveMinimum":0.0,"title":"Weight","description":"Weight must be a positive number in kg"},"marital_status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Marital Status","default":"single"},"family_history":{"type":"boolean","title":"Family History","default":false}},"type":"object","required":["user_id","age","height","weight"],"title":"UserCreate"},"UserResponse":{"properties":{"user_id":{"type":"string","title":"User Id"},"age":{"type":"integer","exclusiveMinimum":0.0,"title":"Age","description":"Age must be a positive integer"},"height":{"type":"number","exclusiveMinimum":0.0,"title":"Height","description":"Height must be a positive number in cm"},"weight":{"type":"number","exclusiveMinimum":0.0,"title":"Weight","description":"Weight must be a positive number in kg"},"marital_status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Marital Status","default":"single"},"family_history":{"type":"boolean","title":"Family History","default":false},"bmi":{"type":"number","title":"Bmi"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["user_id","age","height","weight","bmi","created_at","updated_at"],"title":"UserResponse"},"UserResponsesOut":{"properties":{"user_id":{"type":"string","title":"User Id"},"responses":{"items":{"$ref":"#/components/schemas/SingleResponseOut"},"type":"array","title":"Responses"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["user_id","responses"],"title":"UserResponsesOut"},"UserUpdate":{"properties":{"age":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Age"},"height":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Height"},"weight":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Weight"},"marital_status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Marital Status"},"family_history":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Family History"}},"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}